Shift around tests
