+++ Specs/Scene/MetadataTableSpec.js
@@ -1,4 +1,3 @@
-import { defaultValue } from "../../Source/Cesium.js";
 import { FeatureDetection } from "../../Source/Cesium.js";
 import { MetadataTable } from "../../Source/Cesium.js";
 import MetadataTester from "../MetadataTester.js";
@@ -237,6 +236,29 @@ describe("Scene/MetadataTable", function () {
     expect(results.sort()).toEqual(["height", "name"]);
   });
 
+  it("getProperty", function () {
+    var properties = {
+      propertyInt8: {
+        type: "INT8",
+      },
+    };
+
+    var propertyValues = [-128, 10];
+
+    var metadataTable = MetadataTester.createTable({
+      properties: properties,
+      propertyValues: {
+        propertyInt8: propertyValues,
+      },
+    });
+
+    var length = propertyValues.length;
+    for (var i = 0; i < length; ++i) {
+      var value = metadataTable.getProperty(i, "propertyInt8");
+      expect(value).toEqual(propertyValues[i]);
+    }
+  });
+
   it("getProperty returns undefined when there's no properties", function () {
     var metadataTable = new MetadataTable({
       count: 10,
@@ -259,452 +281,6 @@ describe("Scene/MetadataTable", function () {
     });
 
     expect(metadataTable.getProperty(0, "name")).toBeUndefined();
-  });
-
-  function testGetUint64(options) {
-    options = defaultValue(options, defaultValue.EMPTY_OBJECT);
-    var disableBigIntSupport = options.disableBigIntSupport;
-    var disableBigUint64ArraySupport = options.disableBigUint64ArraySupport;
-
-    var originalValues = [
-      BigInt(0), // eslint-disable-line
-      BigInt(10), // eslint-disable-line
-      BigInt("4611686018427387833"), // eslint-disable-line
-      BigInt("18446744073709551615"), // eslint-disable-line
-    ];
-
-    var expectedValues = originalValues;
-
-    if (disableBigUint64ArraySupport && disableBigIntSupport) {
-      // Precision loss is expected if UINT64 is converted to JS numbers
-      expectedValues = [0, 10, 4611686018427388000, 18446744073709552000];
-    }
-
-    var properties = {
-      propertyUint64: {
-        type: "UINT64",
-      },
-    };
-    var propertyValues = {
-      propertyUint64: originalValues,
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      disableBigUint64ArraySupport: disableBigUint64ArraySupport,
-      disableBigIntSupport: disableBigIntSupport,
-    });
-
-    var length = originalValues.length;
-    for (var i = 0; i < length; ++i) {
-      var value = metadataTable.getProperty(i, "propertyUint64");
-      expect(value).toEqual(expectedValues[i]);
-    }
-  }
-
-  function testGetInt64(options) {
-    options = defaultValue(options, defaultValue.EMPTY_OBJECT);
-    var disableBigIntSupport = options.disableBigIntSupport;
-    var disableBigInt64ArraySupport = options.disableBigInt64ArraySupport;
-
-    var originalValues = [
-      BigInt("-9223372036854775808"), // eslint-disable-line
-      BigInt("-4611686018427387833"), // eslint-disable-line
-      BigInt(-10), // eslint-disable-line
-      BigInt(0), // eslint-disable-line
-      BigInt(10), // eslint-disable-line
-      BigInt("4611686018427387833"), // eslint-disable-line
-      BigInt("9223372036854775807"), // eslint-disable-line
-    ];
-
-    var expectedValues = originalValues;
-
-    if (disableBigInt64ArraySupport && disableBigIntSupport) {
-      // Precision loss is expected if INT64 is converted to JS numbers
-      expectedValues = [
-        -9223372036854776000,
-        -4611686018427388000,
-        -10,
-        0,
-        10,
-        4611686018427388000,
-        9223372036854776000,
-      ];
-    }
-
-    var properties = {
-      propertyInt64: {
-        type: "INT64",
-      },
-    };
-    var propertyValues = {
-      propertyInt64: originalValues,
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      disableBigInt64ArraySupport: disableBigInt64ArraySupport,
-      disableBigIntSupport: disableBigIntSupport,
-    });
-
-    var length = originalValues.length;
-    for (var i = 0; i < length; ++i) {
-      var value = metadataTable.getProperty(i, "propertyInt64");
-      expect(value).toEqual(expectedValues[i]);
-    }
-  }
-
-  it("getProperty returns UINT64 property as BigInt when BigUint64Array is not supported and BigInt is supported", function () {
-    testGetUint64({
-      disableBigUint64ArraySupport: true,
-    });
-  });
-
-  it("getProperty returns UINT64 property as number when BigUint64Array is not supported and BigInt is not supported", function () {
-    testGetUint64({
-      disableBigUint64ArraySupport: true,
-      disableBigIntSupport: true,
-    });
-  });
-
-  it("getProperty returns INT64 property as BigInt when BigInt64Array is supported and BigInt is supported", function () {
-    testGetInt64();
-  });
-
-  it("getProperty returns INT64 property as BigInt when BigInt64Array is not supported and BigInt is supported", function () {
-    testGetInt64({
-      disableBigInt64ArraySupport: true,
-    });
-  });
-
-  it("getProperty returns INT64 property as number when BigInt64Array is not supported and BigInt is not supported", function () {
-    testGetInt64({
-      disableBigInt64ArraySupport: true,
-      disableBigIntSupport: true,
-    });
-  });
-
-  it("getProperty returns single values", function () {
-    // INT64 and UINT64 are tested above
-    var properties = {
-      propertyInt8: {
-        type: "INT8",
-      },
-      propertyUint8: {
-        type: "UINT8",
-      },
-      propertyInt16: {
-        type: "INT16",
-      },
-      propertyUint16: {
-        type: "UINT16",
-      },
-      propertyInt32: {
-        type: "INT32",
-      },
-      propertyUint32: {
-        type: "UINT32",
-      },
-      propertyFloat32: {
-        type: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "FLOAT64",
-      },
-      propertyBoolean: {
-        type: "BOOLEAN",
-      },
-      propertyString: {
-        type: "STRING",
-      },
-      propertyEnum: {
-        type: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [-128, -10, 0, 10, 127],
-      propertyUint8: [0, 10, 20, 30, 255],
-      propertyInt16: [-32768, -10, 0, 10, 32767],
-      propertyUint16: [0, 10, 20, 30, 65535],
-      propertyInt32: [-2147483648, -10, 0, 10, 2147483647],
-      propertyUint32: [0, 10, 20, 30, 4294967295],
-      propertyFloat32: [-2.5, -1.0, 0.0, 700.0, Number.POSITIVE_INFINITY],
-      propertyFloat64: [-234934.12, -1.0, 0.0, 700.0, Number.POSITIVE_INFINITY],
-      propertyBoolean: [false, true, false, true, false],
-      propertyString: ["おはようございます。😃", "a", "", "def", "0001"],
-      propertyEnum: ["ValueA", "ValueB", "Other", "ValueA", "ValueA"],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = propertyValues[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
-  });
-
-  it("getProperty returns fixed size arrays", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-        componentCount: 3,
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-        componentCount: 3,
-      },
-      propertyInt16: {
-        type: "ARRAY",
-        componentType: "INT16",
-        componentCount: 3,
-      },
-      propertyUint16: {
-        type: "ARRAY",
-        componentType: "UINT16",
-        componentCount: 3,
-      },
-      propertyInt32: {
-        type: "ARRAY",
-        componentType: "INT32",
-        componentCount: 3,
-      },
-      propertyUint32: {
-        type: "ARRAY",
-        componentType: "UINT32",
-        componentCount: 3,
-      },
-      propertyInt64: {
-        type: "ARRAY",
-        componentType: "INT64",
-        componentCount: 3,
-      },
-      propertyUint64: {
-        type: "ARRAY",
-        componentType: "UINT64",
-        componentCount: 3,
-      },
-      propertyFloat32: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-        componentCount: 3,
-      },
-      propertyFloat64: {
-        type: "ARRAY",
-        componentType: "FLOAT64",
-        componentCount: 3,
-      },
-      propertyBoolean: {
-        type: "ARRAY",
-        componentType: "BOOLEAN",
-        componentCount: 3,
-      },
-      propertyString: {
-        type: "ARRAY",
-        componentType: "STRING",
-        componentCount: 3,
-      },
-      propertyEnum: {
-        type: "ARRAY",
-        componentType: "ENUM",
-        enumType: "myEnum",
-        componentCount: 3,
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [
-        [-2, -1, 0],
-        [1, 2, 3],
-      ],
-      propertyUint8: [
-        [0, 1, 2],
-        [3, 4, 5],
-      ],
-      propertyInt16: [
-        [-2, -1, 0],
-        [1, 2, 3],
-      ],
-      propertyUint16: [
-        [0, 1, 2],
-        [3, 4, 5],
-      ],
-      propertyInt32: [
-        [-2, -1, 0],
-        [1, 2, 3],
-      ],
-      propertyUint32: [
-        [0, 1, 2],
-        [3, 4, 5],
-      ],
-      propertyInt64: [
-        [BigInt(-2), BigInt(-1), BigInt(0)], // eslint-disable-line
-        [BigInt(1), BigInt(2), BigInt(3)], // eslint-disable-line
-      ],
-      propertyUint64: [
-        [BigInt(0), BigInt(1), BigInt(2)], // eslint-disable-line
-        [BigInt(3), BigInt(4), BigInt(5)], // eslint-disable-line
-      ],
-      propertyFloat32: [
-        [-2.0, -1.0, 0.0],
-        [1.0, 2.0, 3.0],
-      ],
-      propertyFloat64: [
-        [-2.0, -1.0, 0.0],
-        [1.0, 2.0, 3.0],
-      ],
-      propertyBoolean: [
-        [false, true, false],
-        [true, false, true],
-      ],
-      propertyString: [
-        ["a", "bc", "def"],
-        ["dog", "cat", "😃😃😃"],
-      ],
-      propertyEnum: [
-        ["ValueA", "ValueB", "Other"],
-        ["ValueA", "ValueA", "ValueA"],
-      ],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = propertyValues[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
-  });
-
-  it("getProperty returns variable size arrays", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-      },
-      propertyInt16: {
-        type: "ARRAY",
-        componentType: "INT16",
-      },
-      propertyUint16: {
-        type: "ARRAY",
-        componentType: "UINT16",
-      },
-      propertyInt32: {
-        type: "ARRAY",
-        componentType: "INT32",
-      },
-      propertyUint32: {
-        type: "ARRAY",
-        componentType: "UINT32",
-      },
-      propertyInt64: {
-        type: "ARRAY",
-        componentType: "INT64",
-      },
-      propertyUint64: {
-        type: "ARRAY",
-        componentType: "UINT64",
-      },
-      propertyFloat32: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "ARRAY",
-        componentType: "FLOAT64",
-      },
-      propertyBoolean: {
-        type: "ARRAY",
-        componentType: "BOOLEAN",
-      },
-      propertyString: {
-        type: "ARRAY",
-        componentType: "STRING",
-      },
-      propertyEnum: {
-        type: "ARRAY",
-        componentType: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-
-    // Tests empty arrays as well
-    var propertyValues = {
-      propertyInt8: [[-2], [-1, 0], [1, 2, 3], []],
-      propertyUint8: [[0], [1, 2], [3, 4, 5], []],
-      propertyInt16: [[-2], [-1, 0], [1, 2, 3], []],
-      propertyUint16: [[0], [1, 2], [3, 4, 5], []],
-      propertyInt32: [[-2], [-1, 0], [1, 2, 3], []],
-      propertyUint32: [[0], [1, 2], [3, 4, 5], []],
-      propertyInt64: [
-        [BigInt(-2)], // eslint-disable-line
-        [BigInt(-1), BigInt(0)], // eslint-disable-line
-        [BigInt(1), BigInt(2), BigInt(3)], // eslint-disable-line
-        [],
-      ],
-      propertyUint64: [
-        [BigInt(0)], // eslint-disable-line
-        [BigInt(1), BigInt(2)], // eslint-disable-line
-        [BigInt(3), BigInt(4), BigInt(5)], // eslint-disable-line
-        [],
-      ],
-      propertyFloat32: [[-2.0], [-1.0, 0.0], [1.0, 2.0, 3.0], []],
-      propertyFloat64: [[-2.0], [-1.0, 0.0], [1.0, 2.0, 3.0], []],
-      propertyBoolean: [[false], [true, false], [true, false, true], []],
-      propertyString: [["a"], ["bc", "def"], ["dog", "cat", "😃😃😃"], []],
-      propertyEnum: [
-        ["ValueA"],
-        ["ValueB", "Other"],
-        ["ValueA", "ValueA", "ValueA"],
-        [],
-      ],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = propertyValues[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
   });
 
   it("getProperty returns the default value when the property is missing", function () {
@@ -808,6 +384,35 @@ describe("Scene/MetadataTable", function () {
     }).toThrowDeveloperError();
   });
 
+  it("setProperty", function () {
+    var properties = {
+      propertyInt8: {
+        type: "INT8",
+      },
+    };
+
+    var propertyValues = [0, 0];
+    var valuesToSet = [-128, 10];
+
+    var metadataTable = MetadataTester.createTable({
+      properties: properties,
+      propertyValues: {
+        propertyInt8: propertyValues,
+      },
+    });
+
+    var length = valuesToSet.length;
+    for (var i = 0; i < length; ++i) {
+      metadataTable.setProperty(i, "propertyInt8", valuesToSet[i]);
+      var value = metadataTable.getProperty(i, "propertyInt8");
+      expect(value).toEqual(valuesToSet[i]);
+      // Test setting / getting again
+      metadataTable.setProperty(i, "propertyInt8", valuesToSet[i]);
+      value = metadataTable.getProperty(i, "propertyInt8");
+      expect(value).toEqual(valuesToSet[i]);
+    }
+  });
+
   it("setProperty doesn't set property value when there's no properties", function () {
     var metadataTable = new MetadataTable({
       count: 10,
@@ -833,691 +438,6 @@ describe("Scene/MetadataTable", function () {
 
     metadataTable.setProperty(0, "name", "A");
     expect(metadataTable.getProperty(0, "name")).toBeUndefined();
-  });
-
-  it("setProperty sets single values", function () {
-    var properties = {
-      propertyInt8: {
-        type: "INT8",
-      },
-      propertyUint8: {
-        type: "UINT8",
-      },
-      propertyInt16: {
-        type: "INT16",
-      },
-      propertyUint16: {
-        type: "UINT16",
-      },
-      propertyInt32: {
-        type: "INT32",
-      },
-      propertyUint32: {
-        type: "UINT32",
-      },
-      propertyInt64: {
-        type: "INT64",
-      },
-      propertyUint64: {
-        type: "UINT64",
-      },
-      propertyFloat32: {
-        type: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "FLOAT64",
-      },
-      propertyBoolean: {
-        type: "BOOLEAN",
-      },
-      propertyString: {
-        type: "STRING",
-      },
-      propertyEnum: {
-        type: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [0, 0, 0, 0, 0],
-      propertyUint8: [0, 0, 0, 0, 0],
-      propertyInt16: [0, 0, 0, 0, 0],
-      propertyUint16: [0, 0, 0, 0, 0],
-      propertyInt32: [0, 0, 0, 0, 0],
-      propertyUint32: [0, 0, 0, 0, 0],
-      propertyInt64: [BigInt(0), BigInt(0), BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-      propertyUint64: [BigInt(0), BigInt(0), BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-      propertyFloat32: [0.0, 0.0, 0.0, 0.0, 0.0],
-      propertyFloat64: [0.0, 0.0, 0.0, 0.0, 0.0],
-      propertyBoolean: [false, false, true, false, true],
-      propertyString: ["", "", "", "", ""],
-      propertyEnum: ["Other", "Other", "Other", "Other", "Other"],
-    };
-
-    var valuesToSet = {
-      propertyInt8: [-128, -10, 0, 10, 127],
-      propertyUint8: [0, 10, 20, 30, 255],
-      propertyInt16: [-32768, -10, 0, 10, 32767],
-      propertyUint16: [0, 10, 20, 30, 65535],
-      propertyInt32: [-2147483648, -10, 0, 10, 2147483647],
-      propertyUint32: [0, 10, 20, 30, 4294967295],
-      propertyInt64: [
-        BigInt("-9223372036854775808"), // eslint-disable-line
-        BigInt("-4611686018427387833"), // eslint-disable-line
-        BigInt(0), // eslint-disable-line
-        BigInt("4611686018427387833"), // eslint-disable-line
-        BigInt("9223372036854775807"), // eslint-disable-line
-      ],
-      propertyUint64: [
-        BigInt(0), // eslint-disable-line
-        BigInt(10), // eslint-disable-line
-        BigInt(100), // eslint-disable-line
-        BigInt("4611686018427387833"), // eslint-disable-line
-        BigInt("18446744073709551615"), // eslint-disable-line
-      ],
-      propertyFloat32: [-2.5, -1.0, 0.0, 700.0, Number.POSITIVE_INFINITY],
-      propertyFloat64: [-234934.12, -1.0, 0.0, 700.0, Number.POSITIVE_INFINITY],
-      propertyBoolean: [true, true, false, false, true],
-      propertyString: ["おはようございます。😃", "a", "", "def", "0001"],
-      propertyEnum: ["ValueA", "ValueB", "Other", "ValueA", "ValueA"],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = valuesToSet[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-          // Test setting / getting again
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
-  });
-
-  it("setProperty sets fixed size arrays", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-        componentCount: 3,
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-        componentCount: 3,
-      },
-      propertyInt16: {
-        type: "ARRAY",
-        componentType: "INT16",
-        componentCount: 3,
-      },
-      propertyUint16: {
-        type: "ARRAY",
-        componentType: "UINT16",
-        componentCount: 3,
-      },
-      propertyInt32: {
-        type: "ARRAY",
-        componentType: "INT32",
-        componentCount: 3,
-      },
-      propertyUint32: {
-        type: "ARRAY",
-        componentType: "UINT32",
-        componentCount: 3,
-      },
-      propertyInt64: {
-        type: "ARRAY",
-        componentType: "INT64",
-        componentCount: 3,
-      },
-      propertyUint64: {
-        type: "ARRAY",
-        componentType: "UINT64",
-        componentCount: 3,
-      },
-      propertyFloat32: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-        componentCount: 3,
-      },
-      propertyFloat64: {
-        type: "ARRAY",
-        componentType: "FLOAT64",
-        componentCount: 3,
-      },
-      propertyBoolean: {
-        type: "ARRAY",
-        componentType: "BOOLEAN",
-        componentCount: 3,
-      },
-      propertyString: {
-        type: "ARRAY",
-        componentType: "STRING",
-        componentCount: 3,
-      },
-      propertyEnum: {
-        type: "ARRAY",
-        componentType: "ENUM",
-        enumType: "myEnum",
-        componentCount: 3,
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [
-        [0, 0, 0],
-        [0, 0, 0],
-      ],
-      propertyUint8: [
-        [0, 0, 0],
-        [0, 0, 0],
-      ],
-      propertyInt16: [
-        [0, 0, 0],
-        [0, 0, 0],
-      ],
-      propertyUint16: [
-        [0, 0, 0],
-        [0, 0, 0],
-      ],
-      propertyInt32: [
-        [0, 0, 0],
-        [0, 0, 0],
-      ],
-      propertyUint32: [
-        [0, 0, 0],
-        [0, 0, 0],
-      ],
-      propertyInt64: [
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-      ],
-      propertyUint64: [
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-      ],
-      propertyFloat32: [
-        [0.0, 0.0, 0.0],
-        [0.0, 0.0, 0.0],
-      ],
-      propertyFloat64: [
-        [0.0, 0.0, 0.0],
-        [0.0, 0.0, 0.0],
-      ],
-      propertyBoolean: [
-        [false, false, false],
-        [true, true, true],
-      ],
-      propertyString: [
-        ["", "", ""],
-        ["", "", ""],
-      ],
-      propertyEnum: [
-        ["Other", "Other", "Other"],
-        ["Other", "Other", "Other"],
-      ],
-    };
-
-    var valuesToSet = {
-      propertyInt8: [
-        [-2, -1, 0],
-        [1, 2, 3],
-      ],
-      propertyUint8: [
-        [0, 1, 2],
-        [3, 4, 5],
-      ],
-      propertyInt16: [
-        [-2, -1, 0],
-        [1, 2, 3],
-      ],
-      propertyUint16: [
-        [0, 1, 2],
-        [3, 4, 5],
-      ],
-      propertyInt32: [
-        [-2, -1, 0],
-        [1, 2, 3],
-      ],
-      propertyUint32: [
-        [0, 1, 2],
-        [3, 4, 5],
-      ],
-      propertyInt64: [
-        [BigInt(-2), BigInt(-1), BigInt(0)], // eslint-disable-line
-        [BigInt(1), BigInt(2), BigInt(3)], // eslint-disable-line
-      ],
-      propertyUint64: [
-        [BigInt(0), BigInt(1), BigInt(2)], // eslint-disable-line
-        [BigInt(3), BigInt(4), BigInt(5)], // eslint-disable-line
-      ],
-      propertyFloat32: [
-        [-2.0, -1.0, 0.0],
-        [1.0, 2.0, 3.0],
-      ],
-      propertyFloat64: [
-        [-2.0, -1.0, 0.0],
-        [1.0, 2.0, 3.0],
-      ],
-      propertyBoolean: [
-        [false, true, false],
-        [true, false, true],
-      ],
-      propertyString: [
-        ["a", "bc", "def"],
-        ["dog", "cat", "😃😃😃"],
-      ],
-      propertyEnum: [
-        ["ValueA", "ValueB", "Other"],
-        ["ValueA", "ValueA", "ValueA"],
-      ],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = valuesToSet[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-          // Test setting / getting again
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
-  });
-
-  it("setProperty sets variable size arrays with arrays of the same length", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-      },
-      propertyInt16: {
-        type: "ARRAY",
-        componentType: "INT16",
-      },
-      propertyUint16: {
-        type: "ARRAY",
-        componentType: "UINT16",
-      },
-      propertyInt32: {
-        type: "ARRAY",
-        componentType: "INT32",
-      },
-      propertyUint32: {
-        type: "ARRAY",
-        componentType: "UINT32",
-      },
-      propertyInt64: {
-        type: "ARRAY",
-        componentType: "INT64",
-      },
-      propertyUint64: {
-        type: "ARRAY",
-        componentType: "UINT64",
-      },
-      propertyFloat32: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "ARRAY",
-        componentType: "FLOAT64",
-      },
-      propertyBoolean: {
-        type: "ARRAY",
-        componentType: "BOOLEAN",
-      },
-      propertyString: {
-        type: "ARRAY",
-        componentType: "STRING",
-      },
-      propertyEnum: {
-        type: "ARRAY",
-        componentType: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [[0], [0, 0], [0, 0, 0], []],
-      propertyUint8: [[0], [0, 0], [0, 0, 0], []],
-      propertyInt16: [[0], [0, 0], [0, 0, 0], []],
-      propertyUint16: [[0], [0, 0], [0, 0, 0], []],
-      propertyInt32: [[0], [0, 0], [0, 0, 0], []],
-      propertyUint32: [[0], [0, 0], [0, 0, 0], []],
-      propertyInt64: [
-        [BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-        [],
-      ],
-      propertyUint64: [
-        [BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-        [],
-      ],
-      propertyFloat32: [[0.0], [0.0, 0.0], [0.0, 0.0, 0.0], []],
-      propertyFloat64: [[0.0], [0.0, 0.0], [0.0, 0.0, 0.0], []],
-      propertyBoolean: [[false], [false, false], [false, false, false], []],
-      propertyString: [[""], ["", ""], ["", "", ""], []],
-      propertyEnum: [
-        ["Other"],
-        ["Other", "Other"],
-        ["Other", "Other", "Other"],
-        [],
-      ],
-    };
-
-    var valuesToSet = {
-      propertyInt8: [[-2], [-1, 0], [1, 2, 3], []],
-      propertyUint8: [[0], [1, 2], [3, 4, 5], []],
-      propertyInt16: [[-2], [-1, 0], [1, 2, 3], []],
-      propertyUint16: [[0], [1, 2], [3, 4, 5], []],
-      propertyInt32: [[-2], [-1, 0], [1, 2, 3], []],
-      propertyUint32: [[0], [1, 2], [3, 4, 5], []],
-      propertyInt64: [
-        [BigInt(-2)], // eslint-disable-line
-        [BigInt(-1), BigInt(0)], // eslint-disable-line
-        [BigInt(1), BigInt(2), BigInt(3)], // eslint-disable-line
-        [],
-      ],
-      propertyUint64: [
-        [BigInt(0)], // eslint-disable-line
-        [BigInt(1), BigInt(2)], // eslint-disable-line
-        [BigInt(3), BigInt(4), BigInt(5)], // eslint-disable-line
-        [],
-      ],
-      propertyFloat32: [[-2.0], [-1.0, 0.0], [1.0, 2.0, 3.0], []],
-      propertyFloat64: [[-2.0], [-1.0, 0.0], [1.0, 2.0, 3.0], []],
-      propertyBoolean: [[false], [true, false], [true, false, true], []],
-      propertyString: [["a"], ["bc", "def"], ["dog", "cat", "😃😃😃"], []],
-      propertyEnum: [
-        ["ValueA"],
-        ["ValueB", "Other"],
-        ["ValueA", "ValueA", "ValueA"],
-        [],
-      ],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = valuesToSet[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-          // Test setting / getting again
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
-  });
-
-  it("setProperty sets variable size arrays with arrays of different lengths", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-      },
-      propertyInt16: {
-        type: "ARRAY",
-        componentType: "INT16",
-      },
-      propertyUint16: {
-        type: "ARRAY",
-        componentType: "UINT16",
-      },
-      propertyInt32: {
-        type: "ARRAY",
-        componentType: "INT32",
-      },
-      propertyUint32: {
-        type: "ARRAY",
-        componentType: "UINT32",
-      },
-      propertyInt64: {
-        type: "ARRAY",
-        componentType: "INT64",
-      },
-      propertyUint64: {
-        type: "ARRAY",
-        componentType: "UINT64",
-      },
-      propertyFloat32: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "ARRAY",
-        componentType: "FLOAT64",
-      },
-      propertyBoolean: {
-        type: "ARRAY",
-        componentType: "BOOLEAN",
-      },
-      propertyString: {
-        type: "ARRAY",
-        componentType: "STRING",
-      },
-      propertyEnum: {
-        type: "ARRAY",
-        componentType: "ENUM",
-        enumType: "myEnum",
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [[0], [0, 0], [0, 0, 0], []],
-      propertyUint8: [[0], [0, 0], [0, 0, 0], []],
-      propertyInt16: [[0], [0, 0], [0, 0, 0], []],
-      propertyUint16: [[0], [0, 0], [0, 0, 0], []],
-      propertyInt32: [[0], [0, 0], [0, 0, 0], []],
-      propertyUint32: [[0], [0, 0], [0, 0, 0], []],
-      propertyInt64: [
-        [BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-        [],
-      ],
-      propertyUint64: [
-        [BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0)], // eslint-disable-line
-        [BigInt(0), BigInt(0), BigInt(0)], // eslint-disable-line
-        [],
-      ],
-      propertyFloat32: [[0.0], [0.0, 0.0], [0.0, 0.0, 0.0], []],
-      propertyFloat64: [[0.0], [0.0, 0.0], [0.0, 0.0, 0.0], []],
-      propertyBoolean: [[false], [false, false], [false, false, false], []],
-      propertyString: [[""], ["", ""], ["", "", ""], []],
-      propertyEnum: [
-        ["Other"],
-        ["Other", "Other"],
-        ["Other", "Other", "Other"],
-        [],
-      ],
-    };
-
-    var valuesToSet = {
-      propertyInt8: [[1, 2, 3], [], [-2], [-1, 0]],
-      propertyUint8: [[3, 4, 5], [0], [], [1, 2]],
-      propertyInt16: [[], [1, 2, 3], [-2], [-1, 0]],
-      propertyUint16: [[3, 4, 5], [1, 2], [], [0]],
-      propertyInt32: [[1, 2, 3], [], [-2], [-1, 0]],
-      propertyUint32: [[0], [3, 4, 5], [1, 2], []],
-      propertyInt64: [
-        [BigInt(-1), BigInt(0)], // eslint-disable-line
-        [BigInt(-2)], // eslint-disable-line
-        [],
-        [BigInt(1), BigInt(2), BigInt(3)], // eslint-disable-line
-      ],
-      propertyUint64: [
-        [BigInt(0)], // eslint-disable-line
-        [],
-        [BigInt(1), BigInt(2)], // eslint-disable-line
-        [BigInt(3), BigInt(4), BigInt(5)], // eslint-disable-line
-      ],
-      propertyFloat32: [[-1.0, 0.0], [1.0, 2.0, 3.0], [], [-2.0]],
-      propertyFloat64: [[-2.0], [1.0, 2.0, 3.0], [-1.0, 0.0], []],
-      propertyBoolean: [[true, false, true], [], [false], [true, false]],
-      propertyString: [[], ["bc", "def"], ["a"], ["dog", "cat", "😃😃😃"]],
-      propertyEnum: [
-        [],
-        ["ValueA", "ValueA", "ValueA"],
-        ["ValueA"],
-        ["ValueB", "Other"],
-      ],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var expectedValues = valuesToSet[propertyId];
-        var length = expectedValues.length;
-        for (var i = 0; i < length; ++i) {
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-          // Test setting / getting again
-          metadataTable.setProperty(i, propertyId, expectedValues[i]);
-          value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(expectedValues[i]);
-        }
-      }
-    }
-  });
-
-  it("setProperty throws if value doesn't conform to the class property", function () {
-    var properties = {
-      position: {
-        type: "ARRAY",
-        componentType: "FLOAT32",
-        componentCount: 3,
-      },
-    };
-    var propertyValues = {
-      position: [[1.0, 2.0, 3.0]],
-    };
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    expect(function () {
-      metadataTable.setProperty(0, "position", 8.0);
-    }).toThrowDeveloperError();
-  });
-
-  it("setProperty sets Infinity for FLOAT32 and FLOAT64", function () {
-    var properties = {
-      propertyFloat32: {
-        type: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "FLOAT64",
-      },
-    };
-
-    var propertyValues = {
-      propertyFloat32: [0.0, 0.0],
-      propertyFloat64: [0.0, 0.0],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    metadataTable.setProperty(0, "propertyFloat32", Number.POSITIVE_INFINITY);
-    metadataTable.setProperty(1, "propertyFloat32", Number.NEGATIVE_INFINITY);
-    metadataTable.setProperty(0, "propertyFloat64", Number.POSITIVE_INFINITY);
-    metadataTable.setProperty(1, "propertyFloat64", Number.NEGATIVE_INFINITY);
-
-    expect(metadataTable.getProperty(0, "propertyFloat32")).toBe(
-      Number.POSITIVE_INFINITY
-    );
-    expect(metadataTable.getProperty(1, "propertyFloat32")).toBe(
-      Number.NEGATIVE_INFINITY
-    );
-    expect(metadataTable.getProperty(0, "propertyFloat64")).toBe(
-      Number.POSITIVE_INFINITY
-    );
-    expect(metadataTable.getProperty(1, "propertyFloat64")).toBe(
-      Number.NEGATIVE_INFINITY
-    );
-  });
-
-  it("setProperty sets NaN for FLOAT32 and FLOAT64", function () {
-    var properties = {
-      propertyFloat32: {
-        type: "FLOAT32",
-      },
-      propertyFloat64: {
-        type: "FLOAT64",
-      },
-    };
-
-    var propertyValues = {
-      propertyFloat32: [10.0],
-      propertyFloat64: [10.0],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    metadataTable.setProperty(0, "propertyFloat32", NaN);
-    metadataTable.setProperty(0, "propertyFloat64", NaN);
-
-    expect(isNaN(metadataTable.getProperty(0, "propertyFloat32"))).toBe(true);
-    expect(isNaN(metadataTable.getProperty(0, "propertyFloat64"))).toBe(true);
   });
 
   it("setProperty throws without index", function () {
@@ -1600,232 +520,6 @@ describe("Scene/MetadataTable", function () {
 
     expect(function () {
       metadataTable.setProperty(2, "height", 0.0);
-    }).toThrowDeveloperError();
-  });
-
-  it("gets and sets normalized values", function () {
-    var properties = {
-      propertyInt8: {
-        type: "INT8",
-        normalized: true,
-      },
-      propertyUint8: {
-        type: "UINT8",
-        normalized: true,
-      },
-      propertyInt16: {
-        type: "INT16",
-        normalized: true,
-      },
-      propertyUint16: {
-        type: "UINT16",
-        normalized: true,
-      },
-      propertyInt32: {
-        type: "INT32",
-        normalized: true,
-      },
-      propertyUint32: {
-        type: "UINT32",
-        normalized: true,
-      },
-      propertyInt64: {
-        type: "INT64",
-        normalized: true,
-      },
-      propertyUint64: {
-        type: "UINT64",
-        normalized: true,
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [-128, 0, 127],
-      propertyUint8: [0, 51, 255],
-      propertyInt16: [-32768, 0, 32767],
-      propertyUint16: [0, 13107, 65535],
-      propertyInt32: [-2147483648, 0, 2147483647],
-      propertyUint32: [0, 858993459, 4294967295],
-      propertyInt64: [
-        BigInt("-9223372036854775808"), // eslint-disable-line
-        BigInt(0), // eslint-disable-line
-        BigInt("9223372036854775807"), // eslint-disable-line
-      ],
-      propertyUint64: [
-        BigInt(0), // eslint-disable-line
-        BigInt("3689348814741910323"), // eslint-disable-line
-        BigInt("18446744073709551615"), // eslint-disable-line
-      ],
-    };
-
-    var normalizedValues = {
-      propertyInt8: [-1.0, 0, 1.0],
-      propertyUint8: [0.0, 0.2, 1.0],
-      propertyInt16: [-1.0, 0, 1.0],
-      propertyUint16: [0.0, 0.2, 1.0],
-      propertyInt32: [-1.0, 0, 1.0],
-      propertyUint32: [0.0, 0.2, 1.0],
-      propertyInt64: [-1.0, 0, 1.0],
-      propertyUint64: [0.0, 0.2, 1.0],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var length = normalizedValues[propertyId].length;
-        for (var i = 0; i < length; ++i) {
-          var normalizedValue = metadataTable.getProperty(i, propertyId);
-          expect(normalizedValue).toEqual(normalizedValues[propertyId][i]);
-          metadataTable.setProperty(i, propertyId, normalizedValue);
-          normalizedValue = metadataTable.getProperty(i, propertyId);
-          expect(normalizedValue).toEqual(normalizedValues[propertyId][i]);
-        }
-      }
-    }
-  });
-
-  it("gets and sets normalized values for ARRAY type", function () {
-    var properties = {
-      propertyInt8: {
-        type: "ARRAY",
-        componentType: "INT8",
-        normalized: true,
-      },
-      propertyUint8: {
-        type: "ARRAY",
-        componentType: "UINT8",
-        componentCount: 2,
-        normalized: true,
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [[-128, 0], [127], []],
-      propertyUint8: [
-        [0, 255],
-        [0, 51],
-        [255, 255],
-      ],
-    };
-
-    var normalizedValues = {
-      propertyInt8: [[-1.0, 0.0], [1.0], []],
-      propertyUint8: [
-        [0.0, 1.0],
-        [0.0, 0.2],
-        [1.0, 1.0],
-      ],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var length = normalizedValues[propertyId].length;
-        for (var i = 0; i < length; ++i) {
-          var normalizedValue = metadataTable.getProperty(i, propertyId);
-          expect(normalizedValue).toEqual(normalizedValues[propertyId][i]);
-          metadataTable.setProperty(i, propertyId, normalizedValue);
-          normalizedValue = metadataTable.getProperty(i, propertyId);
-          expect(normalizedValue).toEqual(normalizedValues[propertyId][i]);
-        }
-      }
-    }
-  });
-
-  it("does not normalize non integer types", function () {
-    var properties = {
-      propertyEnum: {
-        type: "ENUM",
-        enumType: "myEnum",
-        normalized: true,
-      },
-      propertyEnumArray: {
-        type: "ARRAY",
-        componentType: "ENUM",
-        enumType: "myEnum",
-        normalized: true,
-      },
-      propertyString: {
-        type: "STRING",
-        normalized: true,
-      },
-      propertyBoolean: {
-        type: "BOOLEAN",
-        normalized: true,
-      },
-    };
-
-    var propertyValues = {
-      propertyEnum: ["Other", "ValueA", "ValueB"],
-      propertyEnumArray: [["Other", "ValueA"], ["ValueB"], []],
-      propertyString: ["a", "bc", ""],
-      propertyBoolean: [true, false, false],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-      enums: enums,
-    });
-
-    for (var propertyId in propertyValues) {
-      if (propertyValues.hasOwnProperty(propertyId)) {
-        var length = propertyValues[propertyId].length;
-        for (var i = 0; i < length; ++i) {
-          var value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(propertyValues[propertyId][i]);
-          metadataTable.setProperty(i, propertyId, value);
-          value = metadataTable.getProperty(i, propertyId);
-          expect(value).toEqual(propertyValues[propertyId][i]);
-        }
-      }
-    }
-  });
-
-  it("setProperty throws if value is outside the normalized range", function () {
-    var properties = {
-      propertyInt8: {
-        type: "INT8",
-        normalized: true,
-      },
-      propertyUint8: {
-        type: "UINT8",
-        normalized: true,
-      },
-    };
-
-    var propertyValues = {
-      propertyInt8: [0],
-      propertyUint8: [0],
-    };
-
-    var metadataTable = MetadataTester.createTable({
-      properties: properties,
-      propertyValues: propertyValues,
-    });
-
-    expect(function () {
-      metadataTable.setProperty(0, "propertyInt8", -1.1);
-    }).toThrowDeveloperError();
-
-    expect(function () {
-      metadataTable.setProperty(0, "propertyInt8", 1.1);
-    }).toThrowDeveloperError();
-
-    expect(function () {
-      metadataTable.setProperty(0, "propertyUint8", -0.1);
-    }).toThrowDeveloperError();
-
-    expect(function () {
-      metadataTable.setProperty(0, "propertyUint8", 1.1);
     }).toThrowDeveloperError();
   });
